Split kwarg matcher test into three tests
