+++ test/unit/parameter_matchers/positional_or_keyword_hash_test.rb
@@ -2,7 +2,6 @@
 
 require File.expand_path('../../../test_helper', __FILE__)
 
-require 'deprecation_capture'
 require 'mocha/parameter_matchers/positional_or_keyword_hash'
 require 'mocha/parameter_matchers/instance_methods'
 require 'mocha/inspect'
@@ -11,7 +10,6 @@ require 'mocha/ruby_version'
 
 class PositionalOrKeywordHashTest < Mocha::TestCase
   include Mocha::ParameterMatchers
-  include DeprecationCapture
 
   def test_should_describe_matcher
     hash = { key_1: 1, key_2: 2 }
@@ -61,104 +59,6 @@ class PositionalOrKeywordHashTest < Mocha::TestCase
   def test_should_not_match_keyword_args_with_matchers_using_keyword_args_when_not_all_entries_are_matched
     matcher = build_matcher(Hash.ruby2_keywords_hash({ key_1: is_a(String) }))
     assert !matcher.matches?([Hash.ruby2_keywords_hash({ key_1: 'foo', key_2: 2 })])
-  end
-
-  def test_should_match_hash_arg_with_keyword_args_but_display_deprecation_warning_if_appropriate
-    expectation = Mocha::Expectation.new(self, :foo)
-    matcher = build_matcher(Hash.ruby2_keywords_hash({ key_1: 1, key_2: 2 }), expectation)
-    Mocha::Configuration.override(strict_keyword_argument_matching: false) do
-      capture_deprecation_warnings do
-        assert matcher.matches?([{ key_1: 1, key_2: 2 }])
-      end
-    end
-    return unless Mocha::RUBY_V27_PLUS
-
-    message = last_deprecation_warning
-    location = expectation.definition_location
-    assert_includes message, "Expectation defined at #{location} expected keyword arguments (key_1: 1, key_2: 2)"
-    assert_includes message, 'but received positional hash ({key_1: 1, key_2: 2})'
-    assert_includes message, 'These will stop matching when strict keyword argument matching is enabled.'
-    assert_includes message, 'See the documentation for Mocha::Configuration#strict_keyword_argument_matching=.'
-  end
-
-  def test_should_match_keyword_args_with_hash_arg_but_display_deprecation_warning_if_appropriate
-    expectation = Mocha::Expectation.new(self, :foo)
-    matcher = build_matcher({ key_1: 1, key_2: 2 }, expectation)
-    Mocha::Configuration.override(strict_keyword_argument_matching: false) do
-      capture_deprecation_warnings do
-        assert matcher.matches?([Hash.ruby2_keywords_hash({ key_1: 1, key_2: 2 })])
-      end
-    end
-    return unless Mocha::RUBY_V27_PLUS
-
-    message = last_deprecation_warning
-    location = expectation.definition_location
-    assert_includes message, "Expectation defined at #{location} expected positional hash ({key_1: 1, key_2: 2})"
-    assert_includes message, 'but received keyword arguments (key_1: 1, key_2: 2)'
-    assert_includes message, 'These will stop matching when strict keyword argument matching is enabled.'
-    assert_includes message, 'See the documentation for Mocha::Configuration#strict_keyword_argument_matching=.'
-  end
-
-  if Mocha::RUBY_V27_PLUS
-    def test_should_match_non_last_hash_arg_with_hash_arg_when_strict_keyword_args_is_enabled
-      hash = { key_1: 1, key_2: 2 }
-      matcher = build_matcher(hash)
-      Mocha::Configuration.override(strict_keyword_argument_matching: true) do
-        assert matcher.matches?([{ key_1: 1, key_2: 2 }, %w[a b]])
-      end
-    end
-
-    def test_should_not_match_non_hash_arg_with_hash_arg_when_strict_keyword_args_is_enabled
-      hash = { key_1: 1, key_2: 2 }
-      matcher = build_matcher(hash)
-      Mocha::Configuration.override(strict_keyword_argument_matching: true) do
-        assert !matcher.matches?([%w[a b]])
-      end
-    end
-
-    def test_should_match_hash_arg_with_hash_arg_when_strict_keyword_args_is_enabled
-      hash = { key_1: 1, key_2: 2 }
-      matcher = build_matcher(hash)
-      Mocha::Configuration.override(strict_keyword_argument_matching: true) do
-        assert matcher.matches?([{ key_1: 1, key_2: 2 }])
-      end
-    end
-
-    def test_should_match_keyword_args_with_keyword_args_when_strict_keyword_args_is_enabled
-      matcher = build_matcher(Hash.ruby2_keywords_hash({ key_1: 1, key_2: 2 }))
-      Mocha::Configuration.override(strict_keyword_argument_matching: true) do
-        assert matcher.matches?([Hash.ruby2_keywords_hash({ key_1: 1, key_2: 2 })])
-      end
-    end
-
-    def test_should_not_match_hash_arg_with_keyword_args_when_strict_keyword_args_is_enabled
-      matcher = build_matcher(Hash.ruby2_keywords_hash({ key_1: 1, key_2: 2 }))
-      Mocha::Configuration.override(strict_keyword_argument_matching: true) do
-        assert !matcher.matches?([{ key_1: 1, key_2: 2 }])
-      end
-    end
-
-    def test_should_not_match_keyword_args_with_hash_arg_when_strict_keyword_args_is_enabled
-      hash = { key_1: 1, key_2: 2 }
-      matcher = build_matcher(hash)
-      Mocha::Configuration.override(strict_keyword_argument_matching: true) do
-        assert !matcher.matches?([Hash.ruby2_keywords_hash({ key_1: 1, key_2: 2 })])
-      end
-    end
-
-    def test_should_display_deprecation_warning_even_if_parent_expectation_is_nil
-      expectation = nil
-      matcher = build_matcher({ key_1: 1, key_2: 2 }, expectation)
-      Mocha::Configuration.override(strict_keyword_argument_matching: false) do
-        capture_deprecation_warnings do
-          matcher.matches?([Hash.ruby2_keywords_hash({ key_1: 1, key_2: 2 })])
-        end
-      end
-
-      message = last_deprecation_warning
-      assert_includes message, 'Expectation expected positional hash ({key_1: 1, key_2: 2})'
-      assert_includes message, 'but received keyword arguments (key_1: 1, key_2: 2)'
-    end
   end
 
   private
